+++ changelog.md
@@ -1,1 +1,1 @@
-- Fixed farmer and iron farm not being able to output items
+- Improved universal reputation